feat(address): change address schema to match with database
Previously, address was sent in the backend only with _id for each fields.
But, as updated database schema, we need to send each address fields with
 _id and title. So, client side required the relevant changes.

This commit ensures, every onChange event will change the address state
with the _id and title. To work every thing perfectly as before, every
consumer of this address state also updated.