fix: fix four of address feature's consumer
Change four of address feature's consumer
(EditAddressInfo, EditManualLocationInfo, EditOwnerInfo
DHeroLocation) to adjust with previous breaking changes
in the AddressFields feature. AddressFields isn't requiring
convert addresses to only with _id.

This commit stops addresses converting to get only the _id
of the fields. Because in the previous change, AddressFields
feature requires address state with _id and title for each
fields, which is what exactly coming from the database. So,
converting to only _id is not requiring anymore.